fix "reference delta not found" when cloning from aws codecommit
The existing code does not handle the case where a ref delta refers to another
deltified object.  Once the "parent" was undeltified, not attempt was made to
resolve references to that object's SHA1.

This change simply checks to see if there is an ExternalRef to the object just
resolved, and if so, replaces it and adopts its children.

I'm not an expert in the bowels of git, but this fixed my inability to clone
from codecommit.

Signed-off-by: Kevin Burge <kcburge@gmail.com>